+++ R/srcrefs.R
@@ -163,6 +163,7 @@ srcrefs.nonstandardGenericFunction <- srcrefs.standardGeneric
 
 #' @exportS3Method
 #' @importFrom utils getSrcref
+#' @importFrom methods getPackageName
 #' @rdname srcrefs
 srcrefs.MethodDefinition <- function(x, ..., srcref_names = NULL) {
   # catch methods in methods tables from packages without srcref data
@@ -170,11 +171,11 @@ srcrefs.MethodDefinition <- function(x, ..., srcref_names = NULL) {
 
   # generic source package
   generic_origin_ns <- attr(x@generic, "package")
-  ns <- env_ns_name(environment(x@.Data))
+  ns <- methods::getPackageName(environment(x))
 
   # if method is defined in same package as generic, use generic name as
   # signature alias, otherwise use methods_info-style method alias name
-  if (generic_origin_ns == ns) {
+  if (identical(generic_origin_ns, ns)) {
     signatures <- x@generic
   } else {
     # as produced by `methods:::.methods_info`
@@ -185,7 +186,6 @@ srcrefs.MethodDefinition <- function(x, ..., srcref_names = NULL) {
   objs <- rep_len(list(sr), length(signatures))
 
   names(objs) <- signatures
-  ns <- env_ns_name(environment(x@.Data))
   for (i in seq_along(objs)) attr(objs[[i]], "namespace") <- ns
   objs
 }
@@ -232,9 +232,6 @@ flat_map_srcrefs <- function(xs, ns = NULL, breadcrumbs = character()) {
 
   srcs
 }
-
-
-